Dynamically apply source-maps based on webpack config
Source-maps do not work for CSS when you base64 encode them, yet it
would be nice to to facilitate dynamically applying source-maps using
`webpack -d`. This will apply the sourceMap query param to the
css-loader if a devtool option is selected in the webpack config.

Relates to webpack/less-loader#36
Relates to webpack/css-loader#40