fix - Implement custom `toString` for `extract`
This should fix the behavior with loaders that rely on pitching. Since
it needs to deal with a string, `extract` should return the loader in a
query format as well.

The change allows the `extract` to be used with `loader` field again.

There's one gotcha - the current solution doesn't handle functions
(postcss) in any special way. That may require further thinking.

Closes #364.
Closes #370.