fix(index): id's should be strings (`options.id`)
It currently being an integer causes a bug in case a stringified config
is used for the loader. This means that the plugin instance holds
the ID as an integer, but the loader receives it as string.

The documentation clearly states the ID should be a string, therefor
it should not deviate from that.